Update cat.go & Add out.go: eperate writer from cat.go to out.go and add CloseToDay(), SetCatPrefix()
