+++ cat.go
@@ -1,15 +1,14 @@
 // logcat is a simple logger with color texts in terminal,
 // which just encapsulate golang's log package.
 // and it's eazier for me to use. :p
+// here is basic logger define
 package logcat
 
 import (
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"os"
-	"time"
 )
 
 // csi colors
@@ -22,57 +21,22 @@ const (
 	end     = "\033[0m"
 )
 
-// wrapping the writer with it, when we donot want to see the color stuff.
-type cleanLog struct {
-	file io.Writer
-}
-
-// just strip the '\033[xx m'.
-func (clf cleanLog) Write(b []byte) (n int, err error) {
-	return clf.file.Write(append(b[5:len(b)-5], byte('\n')))
-}
-
 // new logger
 var (
-	goodCat  = log.New(os.Stdout, magenta, log.Ldate|log.Ltime)
-	badCat   = log.New(os.Stdout, red, log.Ldate|log.Ltime)
-	warnCat  = log.New(os.Stdout, yellow, log.Ldate|log.Ltime)
-	infoCat  = log.New(os.Stdout, cyan, log.Ldate|log.Ltime)
-	traceCat = log.New(ioutil.Discard, blue, log.Ldate|log.Ltime)
+	goodCat  = log.New(os.Stdout, magenta, log.LstdFlags)
+	badCat   = log.New(os.Stdout, red, log.LstdFlags)
+	warnCat  = log.New(os.Stdout, yellow, log.LstdFlags)
+	infoCat  = log.New(os.Stdout, cyan, log.LstdFlags)
+	traceCat = log.New(ioutil.Discard, blue, log.LstdFlags)
 )
 
-// return a log file of today, the filename is today's date.
-// prefix is the path to log file. if u leave it empty, it will be ".".
-func Today(prefix string) *os.File {
-	if prefix == "" {
-		prefix = "."
-	}
-	daytime := time.Now().Format("2006-01-02")
-	filename := daytime + ".log"
-	filepath := prefix + "/"
-	os.MkdirAll(filepath, os.ModePerm)
-	logflie, _ := os.OpenFile(filepath+filename, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0644)
-	return logflie
-}
-
-// using a log file to saving bad and warn log. the file could be a writer
-// with Today func, Example: logcat.SetLogFile(logcat.Today("."))
-func SetLogFile(file io.Writer) {
-	if file == os.Stdout {
-		return
-	}
-	wrapFile := cleanLog{file: file}
-	badCat.SetOutput(io.MultiWriter(os.Stdout, wrapFile))
-	warnCat.SetOutput(io.MultiWriter(os.Stdout, wrapFile))
-}
-
-// set the trace out writer
-func SetTraceWay(w io.Writer) {
-	if w == os.Stdout {
-		traceCat.SetOutput(w)
-		return
-	}
-	traceCat.SetOutput(cleanLog{file: w})
+// set the prefix for logger
+func SetCatPrefix(prefix string) {
+	goodCat.SetPrefix(magenta + prefix)
+	badCat.SetPrefix(red + prefix)
+	warnCat.SetPrefix(yellow + prefix)
+	infoCat.SetPrefix(cyan + prefix)
+	traceCat.SetPrefix(blue + prefix)
 }
 
 // format print good log
